pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "elsegate"
version = "0.1.0"
description = "Ollama-compatible LLM gateway. Route requests to multiple backends via YAML config."
readme = "README.md"
license = "AGPL-3.0-or-later"
requires-python = ">=3.11"
authors = [{ name = "Chris Schoemer" }]
keywords = ["ollama", "llm", "gateway", "proxy", "claude", "mistral", "openai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.urls]
Repository = "https://github.com/wizz-cmd/elsegate"

[tool.pytest.ini_options]
asyncio_mode = "auto"
